!!! BUGFIX: Support for `Neos.Flow.cache.applicationIdentifier`
This is helpful to avoid race conditions in an Ansistrano/Surf based
deployment, where the Fusion Cache is shared in a single Redis instance.
In this scenario, make sure to set `Neos.Flow.cache.applicationIdentifier`
in `Settings.yaml` to something release-specific (changing with every deployment).

We need to change the version compatibility of this package because
we rely on the following change - this is also why this change is marked with !!!:

https://github.com/neos/flow-development-collection/pull/2622/commits/98af394ae947c59f851ac260449b293ccfe448b0